+++ src/advanced/__tests__/advanced.test.tsx
@@ -209,7 +209,7 @@ describe("advanced > ", () => {
       expect(screen.queryByText("10개 이상 구매 시 10% 할인")).not.toBeInTheDocument();
       expect(screen.queryByText("5개 이상 구매 시 5% 할인")).toBeInTheDocument();
 
-      fireEvent.click(screen.getAllByText("삭제")[0]);
+      fireEvent.click(screen.getAllByText("삭제")[1]);
       expect(screen.queryByText("10개 이상 구매 시 10% 할인")).not.toBeInTheDocument();
       expect(screen.queryByText("5개 이상 구매 시 5% 할인")).not.toBeInTheDocument();
